Adds support for bridged networking.
The user may create bridges and connect them to host interfaces. Then,
they can specify a bridge with the --bridge option, and qemu will create
a tap device and connect it to that bridge. You can pass the same bridge
multiple times to create several guest interfaces on the same bridge.
The startup script requests a DHCP lease on all interfaces.